backend/requirements.txt
pillow
pytesseract
reportlab
sentence-transformers
chromadb
numpy
pandas
faiss-cpu
transformers
regex
flask
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
aiosqlite
python-multipart
pydantic-settings
starlette
reportlab
Jinja2
httpx